Moved Sidebar, SidebarItem and Topbar into respective folders.
